+++ lib/src/valid_list/valid_list.dart
@@ -1,19 +1,21 @@
-import '../typedefs.dart';
+import 'package:meta/meta.dart';
+
 import 'valid_primitive_list_type.dart';
+import '../typedefs.dart';
 
 class ValidList<V> extends ValidPrimitiveListType<ValidList<V>, V> {
-  factory ValidList([
-    List<V> initialList = const [],
+  ValidList(
+    List<V> initialList, {
     Validator<List<V>>? validator,
-  ]) =>
-      ValidList._(initialList, validator);
-
-  ValidList._(List<V> initialList, Validator<List<V>>? validator)
-      : super.initial(initialList, validator);
+  }) : super.initial(
+          initialList,
+          validator: validator,
+        );
 
   ValidList._next(ValidList<V> previous, List<V> nextList)
       : super.constructNext(previous, nextList);
 
   @override
+  @protected
   ValidList<V> buildNext(List<V> nextList) => ValidList._next(this, nextList);
 }