+++ lib/src/valid_value/valid_primitive_value_type.dart
@@ -1,17 +1,19 @@
-import 'package:valid/src/valid_value/valid_value_type.dart';
-
+import 'valid_value_type.dart';
 import '../typedefs.dart';
 
 abstract class ValidPrimitiveValueType<T extends ValidPrimitiveValueType<T, V>,
     V> extends ValidValueType<T, V> {
   ValidPrimitiveValueType.initial(
-    V initialValue, [
+    V? initialValue, {
     Validator<V>? validator,
-  ])  : assert(
+  })  : assert(
           [bool, int, double, String, DateTime].contains(V),
           'The value type <V> must be one of: [bool, int, double, String, DateTime], received <$V>',
         ),
-        super.initial(initialValue);
+        super.initial(
+          initialValue,
+          validator: validator,
+        );
 
   ValidPrimitiveValueType.constructNext(T previous, V nextValue)
       : super.constructNext(previous, nextValue);