+++ lib/src/validators.dart
@@ -1,5 +1,0 @@
-import 'package:valid/src/typedefs.dart';
-
-Validator<List<V>> every<V>(ListItemValidator<V> itemValidator) {
-  return (list) => list.every(itemValidator);
-}