+++ test/valid_test.dart
@@ -1,1 +1,3 @@
+import 'package:valid/valid.dart';
+
 void main() {}