+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.0-nullsafety.0
 
-- Added base abstract classes
+- Added base abstract classes: valid value, enum, list
 
 ## 0.1.0